requirements.txt
pandas
pytest
streamlit
tabulate
openai
python-dotenv
